riscv: mm: fix map_offset data type
Fix the data type of map_offset to allow storing 64-bit
offset on RV64.

Acked-by: Jens Wiklander <jens.wiklander@linaro.org>

+++ core/arch/riscv/include/mm/core_mmu_arch.h
@@ -161,7 +161,7 @@
 
 struct core_mmu_config {
 	unsigned long satp[CFG_TEE_CORE_NB_CORE];
-	uint32_t map_offset;
+	unsigned long map_offset;
 };
 
 struct core_mmu_user_map {
